+++ src/table/hooks/useRowSelect.tsx
@@ -116,7 +116,6 @@ export default function useRowSelect(
     let selectedRowKeys = [...tSelectedRowKeys.value];
     const reRowKey = props.rowKey || 'id';
     const id = get(row, reRowKey);
-    selectedRowDataMap.value.set(id, row);
     const selectedRowIndex = selectedRowKeys.indexOf(id);
     const isExisted = selectedRowIndex !== -1;
     if (selectColumn.value.type === 'multiple') {
@@ -155,7 +154,7 @@ export default function useRowSelect(
       width: col.width || 64,
       className: [tableSelectedClasses.checkCell, col.className],
       cell: (_: typeof h, p: PrimaryTableCellParams<TableRowData>) => renderSelectCell(p),
-      title: col.type === 'multiple' ? getSelectedHeader() : '',
+      title: col.type === 'multiple' ? getSelectedHeader() : col.title,
     };
   }