+++ src/tag/__tests__/index.test.jsx
@@ -1,5 +1,6 @@
 import { mount } from '@vue/test-utils';
-import { vi } from 'vitest';
+import { vi, expect } from 'vitest';
+import { CloseIcon } from 'tdesign-icons-vue-next';
 import Tag from '@/src/tag/index.ts';
 import CheckTag from '@/src/tag/check-tag.tsx';
 
@@ -11,7 +12,7 @@ describe('Tag or CheckTag', () => {
           return <Tag></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag');
     });
     it(':theme:default', () => {
       const wrapper = mount({
@@ -19,7 +20,7 @@ describe('Tag or CheckTag', () => {
           return <Tag theme={'default'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--default');
     });
     it(':theme:primary', () => {
       const wrapper = mount({
@@ -27,7 +28,7 @@ describe('Tag or CheckTag', () => {
           return <Tag theme={'primary'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--primary');
     });
     it(':theme:warning', () => {
       const wrapper = mount({
@@ -35,7 +36,7 @@ describe('Tag or CheckTag', () => {
           return <Tag theme={'warning'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--warning');
     });
     it(':theme:danger', () => {
       const wrapper = mount({
@@ -43,7 +44,7 @@ describe('Tag or CheckTag', () => {
           return <Tag theme={'danger'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--danger');
     });
     it(':theme:success', () => {
       const wrapper = mount({
@@ -51,32 +52,20 @@ describe('Tag or CheckTag', () => {
           return <Tag theme={'success'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--success');
     });
 
     it(':size:small', () => {
-      const wrapper = mount({
-        render() {
-          <Tag size="small"></Tag>;
-        },
-      });
-      expect(wrapper.element).toMatchSnapshot();
+      const wrapper = mount(() => <Tag size="small"></Tag>);
+      expect(wrapper.classes()).toContain('t-size-s');
     });
     it(':size:medium', () => {
-      const wrapper = mount({
-        render() {
-          <Tag size="medium"></Tag>;
-        },
-      });
-      expect(wrapper.element).toMatchSnapshot();
+      const wrapper = mount(() => <Tag size="medium"></Tag>);
+      expect(wrapper.classes()).toContain('t-size-m');
     });
     it(':size:large', () => {
-      const wrapper = mount({
-        render() {
-          return <Tag size={'large'}></Tag>;
-        },
-      });
-      expect(wrapper.element).toMatchSnapshot();
+      const wrapper = mount(() => <Tag size="large"></Tag>);
+      expect(wrapper.classes()).toContain('t-size-l');
     });
 
     it(':closable', () => {
@@ -86,18 +75,9 @@ describe('Tag or CheckTag', () => {
           return <Tag closable onClose={fn}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.findComponent(CloseIcon)).toBeTruthy();
       wrapper.find('.t-icon-close').trigger('click');
       expect(fn).toHaveBeenCalled();
-    });
-
-    it(':checked', () => {
-      const wrapper = mount({
-        render() {
-          return <Tag checked></Tag>;
-        },
-      });
-      expect(wrapper.element).toMatchSnapshot();
     });
 
     it(':disabled', () => {
@@ -107,8 +87,8 @@ describe('Tag or CheckTag', () => {
           return <Tag disabled onClick={fn}></Tag>;
         },
       });
+      expect(wrapper.classes()).toContain('t-tag--disabled');
       expect(fn).not.toHaveBeenCalled();
-      expect(wrapper.element).toMatchSnapshot();
     });
 
     it(':variant:dark;theme:default', () => {
@@ -117,7 +97,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="dark" theme={'default'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--dark');
+      expect(wrapper.classes()).toContain('t-tag--default');
     });
     it(':variant:dark;theme:primary', () => {
       const wrapper = mount({
@@ -125,15 +106,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="dark" theme={'primary'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
-    });
-    it(':variant:dark;theme:default', () => {
-      const wrapper = mount({
-        render() {
-          return <Tag variant="dark" theme={'default'}></Tag>;
-        },
-      });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--dark');
+      expect(wrapper.classes()).toContain('t-tag--primary');
     });
     it(':variant:dark;theme:warning', () => {
       const wrapper = mount({
@@ -141,7 +115,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="dark" theme={'warning'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--dark');
+      expect(wrapper.classes()).toContain('t-tag--warning');
     });
     it(':variant:dark;theme:danger', () => {
       const wrapper = mount({
@@ -149,7 +124,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="dark" theme={'danger'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--dark');
+      expect(wrapper.classes()).toContain('t-tag--danger');
     });
     it(':variant:dark;theme:success', () => {
       const wrapper = mount({
@@ -157,7 +133,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="dark" theme={'success'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--dark');
+      expect(wrapper.classes()).toContain('t-tag--success');
     });
 
     it(':variant:light-outline;theme:default', () => {
@@ -166,7 +143,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="light" theme={'default'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--light');
+      expect(wrapper.classes()).toContain('t-tag--default');
     });
     it(':variant:light-outline;theme:primary', () => {
       const wrapper = mount({
@@ -174,15 +152,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="light" theme={'primary'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
-    });
-    it(':variant:light-outline;theme:default', () => {
-      const wrapper = mount({
-        render() {
-          return <Tag variant="light" theme={'default'}></Tag>;
-        },
-      });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--light');
+      expect(wrapper.classes()).toContain('t-tag--primary');
     });
     it(':variant:light-outline;theme:warning', () => {
       const wrapper = mount({
@@ -190,7 +161,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="light" theme={'warning'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--light');
+      expect(wrapper.classes()).toContain('t-tag--warning');
     });
     it(':variant:light-outline;theme:danger', () => {
       const wrapper = mount({
@@ -198,7 +170,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="light" theme={'danger'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--light');
+      expect(wrapper.classes()).toContain('t-tag--danger');
     });
     it(':variant:light-outline;theme:success', () => {
       const wrapper = mount({
@@ -206,7 +179,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="light" theme={'success'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--light');
+      expect(wrapper.classes()).toContain('t-tag--success');
     });
     it(':variant:light;theme:default', () => {
       const wrapper = mount({
@@ -214,7 +188,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="light" theme={'default'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--light');
+      expect(wrapper.classes()).toContain('t-tag--default');
     });
     it(':variant:light;theme:primary', () => {
       const wrapper = mount({
@@ -222,15 +197,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="light" theme={'primary'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
-    });
-    it(':variant:light;theme:default', () => {
-      const wrapper = mount({
-        render() {
-          return <Tag variant="light" theme={'default'}></Tag>;
-        },
-      });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--light');
+      expect(wrapper.classes()).toContain('t-tag--primary');
     });
     it(':variant:light;theme:warning', () => {
       const wrapper = mount({
@@ -238,7 +206,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="light" theme={'warning'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--light');
+      expect(wrapper.classes()).toContain('t-tag--warning');
     });
     it(':variant:light;theme:danger', () => {
       const wrapper = mount({
@@ -246,7 +215,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="light" theme={'danger'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--light');
+      expect(wrapper.classes()).toContain('t-tag--danger');
     });
     it(':variant:light;theme:success', () => {
       const wrapper = mount({
@@ -254,7 +224,8 @@ describe('Tag or CheckTag', () => {
           return <Tag variant="light" theme={'success'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--light');
+      expect(wrapper.classes()).toContain('t-tag--success');
     });
 
     it(':shape:round', () => {
@@ -263,7 +234,7 @@ describe('Tag or CheckTag', () => {
           return <Tag shape={'round'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--round');
     });
     it(':shape:square', () => {
       const wrapper = mount({
@@ -271,7 +242,7 @@ describe('Tag or CheckTag', () => {
           return <Tag shape={'square'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag');
     });
     it(':shape:mark', () => {
       const wrapper = mount({
@@ -279,7 +250,7 @@ describe('Tag or CheckTag', () => {
           return <Tag shape={'mark'}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--mark');
     });
 
     it(':maxWidth', () => {
@@ -288,7 +259,8 @@ describe('Tag or CheckTag', () => {
           return <Tag max-width={100}></Tag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      const tag = wrapper.find('.t-tag');
+      expect(getComputedStyle(tag.element, null).maxWidth).toBe('100px');
     });
   });
 
@@ -299,7 +271,8 @@ describe('Tag or CheckTag', () => {
           return <CheckTag checked></CheckTag>;
         },
       });
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag');
+      expect(wrapper.classes()).toContain('t-tag--checked');
     });
     it(':disabled', () => {
       const fn = vi.fn();
@@ -309,18 +282,18 @@ describe('Tag or CheckTag', () => {
         },
       });
       expect(fn).not.toHaveBeenCalled();
-      expect(wrapper.element).toMatchSnapshot();
+      expect(wrapper.classes()).toContain('t-tag--disabled');
     });
   });
 
   describe('Tag:slot', () => {
     it('<icon>', () => {
-      const wrapper = mount(Tag, {
-        scopedSlots: {
-          icon: '<div></div>',
-        },
-      });
-      expect(wrapper.element).toMatchSnapshot();
+      const slots = {
+        icon: () => <div>text</div>,
+      };
+      const wrapper = mount(() => <Tag v-slots={slots} />);
+      const tag = wrapper.find('.t-tag div');
+      expect(tag.text()).toBe('text');
     });
   });