+++ src/upload/_example/request-method.vue
@@ -11,7 +11,7 @@
       ref="uploadRef"
       v-model="files"
       :request-method="requestMethod"
-      tips="自定义上传方法需要返回成功或失败信息"
+      placeholder="自定义上传方法需要返回成功或失败信息"
       :on-fail="handleRequestFail"
     ></t-upload>
   </div>
@@ -29,12 +29,6 @@ const requestSuccessMethod = (file) => {
     // file.percent 用于控制上传进度，如果不希望显示上传进度，则不对 file.percent 设置值即可。
     // 如果代码规范不能设置 file.percent，也可以设置 files
     file.percent = 0;
-
-    setTimeout(() => {
-      // 设置上传进度
-      uploadRef.value.setPercent(50);
-    }, 500);
-
     const timer = setTimeout(() => {
       // resolve 参数为关键代码
       resolve({ status: 'success', response: { url: 'https://tdesign.gtimg.com/site/avatar.jpg' } });