+++ src/upload/_example/single-input.vue
@@ -1,14 +1,14 @@
 <template>
   <div class="tdesign-demo-upload">
     <div style="width: 350px">
+      <!-- abridgeName 省略中间文本，首尾保留的文本字符 -->
       <t-upload
         v-model="files"
         action="https://service-bv448zsw-1257786608.gz.apigw.tencentcs.com/api/upload-demo"
-        :tips="tips"
+        :abridge-name="[8, 6]"
         theme="file-input"
         placeholder="未选择文件"
         @fail="handleFail"
-        @success="onSuccess"
       ></t-upload>
     </div>
   </div>
@@ -18,13 +18,8 @@ import { ref } from 'vue';
 import { MessagePlugin } from 'tdesign-vue-next';
 
 const files = ref([]);
-const tips = ref('');
 
-const handleFail = (file) => {
+const handleFail = ({ file }) => {
   MessagePlugin.error(`文件 ${file.name} 上传失败`);
-};
-
-const onSuccess = () => {
-  tips.value = '';
 };
 </script>