+++ src/upload/hooks/useDrag.ts
@@ -1,0 +1,48 @@
+import { ref } from 'vue';
+import { TdUploadProps } from '../type';
+
+export interface UploadDragEvents {
+  onDragFileChange?: (e: DragEvent) => void;
+  onDragenter?: TdUploadProps['onDragenter'];
+  onDragleave?: TdUploadProps['onDragleave'];
+  onDrop?: TdUploadProps['onDrop'];
+}
+
+export default function useDrag(props: UploadDragEvents) {
+  const target = ref(null);
+  const dragActive = ref(false);
+
+  const handleDrop = (event: DragEvent) => {
+    event.preventDefault();
+    props.onDragFileChange?.(event);
+    props.onDrop?.({ e: event });
+    dragActive.value = false;
+  };
+
+  const handleDragenter = (event: DragEvent) => {
+    event.preventDefault();
+    target.value = event.target;
+    props.onDragenter?.({ e: event });
+    dragActive.value = true;
+  };
+
+  const handleDragleave = (event: DragEvent) => {
+    if (event.target !== target.value) return;
+    event.preventDefault();
+    props.onDragleave?.({ e: event });
+    dragActive.value = false;
+  };
+
+  const handleDragover = (event: DragEvent) => {
+    event.preventDefault();
+  };
+
+  return {
+    target,
+    dragActive,
+    handleDrop,
+    handleDragenter,
+    handleDragleave,
+    handleDragover,
+  };
+}